fix: offeredPayment type INT4 not FLOAT8
RescueRow.offered_payment was Option<f64> but the Postgres column is INT4.
Changed to Option<i32> to match the actual schema.

+++ guardian-engine/src/routes/views.rs
@@ -625,7 +625,7 @@ struct RescueRow {
     #[sqlx(rename = "medicalRequired")]
     medical_required: bool,
     #[sqlx(rename = "offeredPayment")]
-    offered_payment: Option<f64>,
+    offered_payment: Option<i32>,
 }
 
 #[derive(sqlx::FromRow)]
